fetchdata using - useReducer ----- imp**** ------
 const [state, dispatch] = useReducer(reducer, initialState)

dispatch ( { type : "FETCH_SUCCESS" , payload  :  data } )

Here, dispatch is a function, and you are dispatching an action by providing an object with a type property (and optionally a payload property). The reducer function, in turn, will handle this action and update the state accordingly.

function  reducer (state, action) { upadate state}

A reducer in React is a function that takes the current state and an action as parameters. It updates the state based on the action type using a switch statement or other conditional logic. The action type, provided by the dispatch function, guides how the state should be modified. The reducer returns a new state without modifying the existing one.